Extract CAR into an object
